requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==44.0.2
python-jose[cryptography]==3.3.0
requests==2.32.3
pytest==8.3.5
pytest-flask==1.3.0
Werkzeug==3.1.3
gunicorn==23.0.0